+++ mail/proton-mail-common/CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased] - 2024-00-00
 
+### Added
+
+- Strip UTM parameters from HTML content
+- [iOS] Inject viewport metadata for web view
+
 ## [0.5.28] - 2024-07-02
 
 ### Fixed